requirements.txt
aniso8601==9.0.1
APScheduler==3.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dataclasses==0.8
decorator==5.0.9
future==0.18.2
greenlet==1.1.0
idna==3.2
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
keyboa==2.2.8
lxml==4.6.3
MarkupSafe==2.0.1
pycparser==2.20
pyTelegramBotAPI==3.8.3
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.20
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.4.1
